Refactor/sprint 2

+++ src/domains/implementation-plan/services/updateImplementationPlan.ts
@@ -1,13 +1,24 @@
 import { prisma } from '@/lib/prisma';
 
 export default async function updateImplementationPlan(serviceRequestId: string, tasks: Task[]) {
-  return await prisma.implementationPlan.update({
-    where: { id: serviceRequestId },
-    data: {
-      tasks: {
-        deleteMany: {},
-        create: tasks,
-      },
-    },
+  const existingPlan = await prisma.implementationPlan.findUnique({
+    where: { serviceRequestId },
   });
+
+  if (existingPlan) {
+    await prisma.implementationPlan.update({
+      where: { id: existingPlan.id },
+      data: {
+        tasks: {
+          deleteMany: {},
+          create: tasks.map(task => ({
+            id: task.id,
+            name: task.name,
+            deadline: task.deadline,
+            checked: task.checked,
+          })),
+        },
+      },
+    });
+  }
 }
